backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pydantic[email]==2.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
langchain-google-genai==2.0.5
pdfplumber==0.11.4
openai-whisper==20240930
faiss-cpu==1.8.0
sentence-transformers==3.0.1
aiofiles==24.1.0
redis==5.0.8
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
